add the richtext editor module
