+++ code/HoneyComb/package.json
@@ -20,6 +20,7 @@
     "node-sass": "^4.9.2",
     "sass-loader": "^7.0.3",
     "vue": "^2.5.2",
+    "vue-quill-editor": "^3.0.6",
     "vue-router": "^3.0.1",
     "vue-scroll": "^2.1.7"
   },
