+++ code/HoneyComb/static/css/markdown.css
@@ -14,7 +14,7 @@
 
 .navContainer {
   width: 100%;
-  height: 36px;
+  height: 42px;
   background: #fff;
   box-sizing: border-box;
   border-bottom: 1px solid #eee;
@@ -44,7 +44,7 @@ li.markListItem {
   list-style: none;
   width: 20px;
   height: 20px;
-  margin: 0 2px;
+  margin: 0 2% 0 2%;
   display: flex;
   justify-content: center;
   align-items: center;
